+++ client/network/Cargo.toml
@@ -67,7 +67,6 @@ libp2p = "0.44.0"
 
 [dev-dependencies]
 assert_matches = "1.3"
-libp2p = { version = "0.44.0", default-features = false }
 quickcheck = "1.0.3"
 rand = "0.7.2"
 sp-test-primitives = { version = "2.0.0", path = "../../primitives/test-primitives" }
