+++ client/network/sync/Cargo.toml
@@ -20,6 +20,7 @@ prost-build = "0.9"
 codec = { package = "parity-scale-codec", version = "3.0.0", features = [
     "derive",
 ] }
+libp2p = "0.44.0"
 log = "0.4.16"
 prost = "0.9"
 sc-client-api = { version = "4.0.0-dev", path = "../../api" }
