+++ client/network/sync/src/lib.rs
@@ -16,5 +16,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
+pub mod blocks;
+pub mod message;
 pub mod schema;
 pub mod state;
